+++ src/client/gui/client_ui.cpp
@@ -68,6 +68,8 @@ mbClientUi::mbClientUi(mbClient *core, QWidget *parent) :
     m_ui.menuHelp                        = ui->menuHelp                       ;
     m_ui.actionFileNew                   = ui->actionFileNew                  ;
     m_ui.actionFileOpen                  = ui->actionFileOpen                 ;
+    m_ui.actionFileRecent                = ui->actionFileRecent               ;
+    m_ui.actionFileClose                 = ui->actionFileClose                ;
     m_ui.actionFileSave                  = ui->actionFileSave                 ;
     m_ui.actionFileSaveAs                = ui->actionFileSaveAs               ;
     m_ui.actionFileEdit                  = ui->actionFileEdit                 ;